+++ tests/testthat/test-family.R
@@ -490,59 +490,55 @@ test_that("sampling works (vpoisson)", {
   W <- generateW("rectangle", 100, 2, 10)
   model_orders <- list(intercept = "homogeneous", past_obs = 2, past_mean = 1)
   parameter <- list(intercept = 0.5, past_obs = matrix(c(0.3, 0.2, 0.05), nrow = 3), 
-                    past_mean = matrix(c(0.1, 0.05), nrow = 2),
-                    covariates = c(0.75, 0.5))
-  covariates <- list(season = SpatialConstant(sin(2* pi / 12 * seq(n_obs))),
-                   location = TimeConstant(rnorm(100, sd = 0.81)))
-  
+                    past_mean = matrix(c(0.1, 0.05), nrow = 2))
+
   # log-link
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("log", sampling_method = "inversion"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("log", sampling_method = "poisson_process"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("log", sampling_method = "inversion"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("log", sampling_method = "poisson_process"))
+  expect_true(is.list(sim))
   
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("log", copula = "frank", copula_param = 2, sampling_method = "inversion"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("log", copula = "frank", copula_param = 2, sampling_method = "poisson_process"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("log", copula = "frank", copula_param = 2, sampling_method = "inversion"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("log", copula = "frank", copula_param = 2, sampling_method = "poisson_process"))
+  expect_true(is.list(sim))
 
   # identity-link
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("identity", sampling_method = "inversion"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("identity", sampling_method = "poisson_process"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("identity", sampling_method = "inversion"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("identity", sampling_method = "poisson_process"))
+  expect_true(is.list(sim))
 
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("identity", copula = "frank", copula_param = 2, sampling_method = "inversion"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("identity", copula = "frank", copula_param = 2, sampling_method = "poisson_process"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("identity", copula = "frank", copula_param = 2, sampling_method = "inversion"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("identity", copula = "frank", copula_param = 2, sampling_method = "poisson_process"))
+  expect_true(is.list(sim))
 
  # root link
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("sqrt", sampling_method = "inversion"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("sqrt", sampling_method = "poisson_process"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("sqrt", sampling_method = "inversion"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("sqrt", sampling_method = "poisson_process"))
+  expect_true(is.list(sim))
 
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("sqrt", copula = "frank", copula_param = 2, sampling_method = "inversion"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("sqrt", copula = "frank", copula_param = 2, sampling_method = "poisson_process"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("sqrt", copula = "frank", copula_param = 2, sampling_method = "inversion"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("sqrt", copula = "frank", copula_param = 2, sampling_method = "poisson_process"))
+  expect_true(is.list(sim))
 
   # softplus link
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("softplus", const = 2, sampling_method = "inversion"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("softplus", const = 2, sampling_method = "poisson_process"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("softplus", const = 2, sampling_method = "inversion"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("softplus", const = 2, sampling_method = "poisson_process"))
+  expect_true(is.list(sim))
 
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("softplus", const = 2, copula = "frank", copula_param = 2, sampling_method = "inversion"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vpoisson("softplus", const = 2, copula = "frank", copula_param = 2, sampling_method = "poisson_process"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("softplus", const = 2, copula = "frank", copula_param = 2, sampling_method = "inversion"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vpoisson("softplus", const = 2, copula = "frank", copula_param = 2, sampling_method = "poisson_process"))
+  expect_true(is.list(sim))
 })
 
 
 
-# sampling tests within C++
 testthat::skip_on_cran()
 test_that("sampling works (vquasipoisson)", {
   set.seed(42)
@@ -550,104 +546,297 @@ test_that("sampling works (vquasipoisson)", {
   W <- generateW("rectangle", 100, 2, 10)
   model_orders <- list(intercept = "homogeneous", past_obs = 2, past_mean = 1)
   parameter <- list(intercept = 0.5, past_obs = matrix(c(0.3, 0.2, 0.05), nrow = 3), 
-                    past_mean = matrix(c(0.1, 0.05), nrow = 2),
-                    covariates = c(0.75, 0.5))
-  covariates <- list(season = SpatialConstant(sin(2* pi / 12 * seq(n_obs))),
-                   location = TimeConstant(rnorm(100, sd = 0.81)))
+                    past_mean = matrix(c(0.1, 0.05), nrow = 2))
   
   # log-link
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("log", dispersion = 2, sampling_method = "build_up"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("log", dispersion = 2, sampling_method = "chop_down"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("log", dispersion = 2, sampling_method = "branching"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("log", dispersion = 2, sampling_method = "negbin"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("log", dispersion = 2, sampling_method = "build_up"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("log", dispersion = 2, sampling_method = "chop_down"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("log", dispersion = 2, sampling_method = "branching"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("log", dispersion = 2, sampling_method = "negbin"))
+  expect_true(is.list(sim))
 
 
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("log", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "build_up"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("log", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "chop_down"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("log", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "branching"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("log", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "negbin"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("log", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "build_up"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("log", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "chop_down"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("log", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "branching"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("log", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "negbin"))
+  expect_true(is.list(sim))
 
 
   # identity-link
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("identity", dispersion = 2, sampling_method = "build_up"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("identity", dispersion = 2, sampling_method = "chop_down"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("identity", dispersion = 2, sampling_method = "branching"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("identity", dispersion = 2, sampling_method = "negbin"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("identity", dispersion = 2, sampling_method = "build_up"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("identity", dispersion = 2, sampling_method = "chop_down"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("identity", dispersion = 2, sampling_method = "branching"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("identity", dispersion = 2, sampling_method = "negbin"))
+  expect_true(is.list(sim))
 
 
 
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("identity", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "build_up"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("identity", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "chop_down"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("identity", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "branching"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("identity", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "negbin"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("identity", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "build_up"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("identity", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "chop_down"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("identity", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "branching"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("identity", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "negbin"))
+  expect_true(is.list(sim))
 
  # root link
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("sqrt", dispersion = 2, sampling_method = "build_up"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("sqrt", dispersion = 2, sampling_method = "chop_down"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("sqrt", dispersion = 2, sampling_method = "branching"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("sqrt", dispersion = 2, sampling_method = "negbin"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("sqrt", dispersion = 2, sampling_method = "build_up"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("sqrt", dispersion = 2, sampling_method = "chop_down"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("sqrt", dispersion = 2, sampling_method = "branching"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("sqrt", dispersion = 2, sampling_method = "negbin"))
+  expect_true(is.list(sim))
 
 
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("sqrt", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "build_up"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("sqrt", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "chop_down"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("sqrt", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "branching"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("sqrt", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "negbin"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("sqrt", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "build_up"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("sqrt", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "chop_down"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("sqrt", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "branching"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("sqrt", dispersion = 2, copula = "frank", copula_param = 2, sampling_method = "negbin"))
+  expect_true(is.list(sim))
 
 
   # softplus link
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("softplus", dispersion = 2, const = 2, sampling_method = "build_up"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("softplus", dispersion = 2, const = 2, sampling_method = "chop_down"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("softplus", dispersion = 2, const = 2, sampling_method = "branching"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("softplus", dispersion = 2, const = 2, sampling_method = "negbin"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("softplus", dispersion = 2, const = 2, sampling_method = "build_up"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("softplus", dispersion = 2, const = 2, sampling_method = "chop_down"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("softplus", dispersion = 2, const = 2, sampling_method = "branching"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("softplus", dispersion = 2, const = 2, sampling_method = "negbin"))
+  expect_true(is.list(sim))
 
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("softplus", dispersion = 2, const = 2, copula = "frank", copula_param = 2, sampling_method = "build_up"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("softplus", dispersion = 2, const = 2, copula = "frank", copula_param = 2, sampling_method = "chop_down"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("softplus", dispersion = 2, const = 2, copula = "frank", copula_param = 2, sampling_method = "branching"))
-  expect_s3_class(sim, "list")
-  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, covariates, family = vquasipoisson("softplus", dispersion = 2, const = 2, copula = "frank", copula_param = 2, sampling_method = "negbin"))
-  expect_s3_class(sim, "list")
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("softplus", dispersion = 2, const = 2, copula = "frank", copula_param = 2, sampling_method = "build_up"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("softplus", dispersion = 2, const = 2, copula = "frank", copula_param = 2, sampling_method = "chop_down"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("softplus", dispersion = 2, const = 2, copula = "frank", copula_param = 2, sampling_method = "branching"))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasipoisson("softplus", dispersion = 2, const = 2, copula = "frank", copula_param = 2, sampling_method = "negbin"))
+  expect_true(is.list(sim))
 })
 
 
 
 
 
+testthat::skip_on_cran()
+test_that("sampling works (vnegative.binomial)", {
+  set.seed(42)
+  n_obs <- 200L
+  W <- generateW("rectangle", 100, 2, 10)
+  model_orders <- list(intercept = "homogeneous", past_obs = 2, past_mean = 1)
+  parameter <- list(intercept = 0.5, past_obs = matrix(c(0.3, 0.2, 0.05), nrow = 3), 
+                    past_mean = matrix(c(0.1, 0.05), nrow = 2))
+  
+  # log-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnegative.binomial("log", dispersion = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnegative.binomial("log", dispersion = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+  # identity-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnegative.binomial("identity", dispersion = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnegative.binomial("identity", dispersion = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+ # root link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnegative.binomial("sqrt", dispersion = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnegative.binomial("sqrt", dispersion = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+  # softplus link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnegative.binomial("softplus", dispersion = 2, const = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnegative.binomial("softplus", dispersion = 2, const = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+})
+
+
+
+
+testthat::skip_on_cran()
+test_that("sampling works (vbinomial)", {
+  set.seed(42)
+  n_obs <- 200L
+  W <- generateW("rectangle", 100, 2, 10)
+  model_orders <- list(intercept = "homogeneous", past_obs = 2, past_mean = 1)
+  parameter <- list(intercept = 0.5, past_obs = matrix(c(0.3, 0.2, 0.05), nrow = 3), 
+                    past_mean = matrix(c(0.1, 0.05), nrow = 2))
+  
+  # log-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vbinomial("softclipping", size = 10, const = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vbinomial("softclipping", size = 10, const = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+  # identity-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vbinomial("identity", size = 10))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vbinomial("identity", size = 10, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+ # root link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vbinomial("logit", size = 10))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vbinomial("logit", size = 10, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+  # softplus link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vbinomial("probit", size = 10))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vbinomial("probit", size = 10, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+})
 
 
 
 
 
+testthat::skip_on_cran()
+test_that("sampling works (vquasibinomial)", {
+  set.seed(42)
+  n_obs <- 200L
+  W <- generateW("rectangle", 100, 2, 10)
+  model_orders <- list(intercept = "homogeneous", past_obs = 2, past_mean = 1)
+  parameter <- list(intercept = 0.5, past_obs = matrix(c(0.3, 0.2, 0.05), nrow = 3), 
+                    past_mean = matrix(c(0.1, 0.05), nrow = 2))
+  
+  # log-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasibinomial("softclipping", size = 10, dispersion = 2, const = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasibinomial("softclipping", size = 10, dispersion = 2, const = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
 
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasibinomial("softclipping", size = 10, dispersion = 0.9, const = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasibinomial("softclipping", size = 10, dispersion = 0.9, const = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+  # identity-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasibinomial("identity", size = 10, dispersion = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasibinomial("identity", size = 10, dispersion = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+ # root link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasibinomial("logit", size = 10, dispersion = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasibinomial("logit", size = 10, dispersion = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+  # softplus link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasibinomial("probit", size = 10, dispersion = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vquasibinomial("probit", size = 10, dispersion = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+})
+
+
+
+testthat::skip_on_cran()
+test_that("sampling works (vgamma)", {
+  set.seed(42)
+  n_obs <- 200L
+  W <- generateW("rectangle", 100, 2, 10)
+  model_orders <- list(intercept = "homogeneous", past_obs = 2, past_mean = 1)
+  parameter <- list(intercept = 0.5, past_obs = matrix(c(0.3, 0.2, 0.05), nrow = 3), 
+                    past_mean = matrix(c(0.1, 0.05), nrow = 2))
+  
+  # inverse-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vgamma("inverse", size = 10, dispersion = 2, const = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vgamma("inverse", size = 10, dispersion = 2, const = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+  # log-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vgamma("log", size = 10, dispersion = 0.9, const = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vgamma("log", size = 10, dispersion = 0.9, const = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+  # identity-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vgamma("identity", size = 10, dispersion = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vgamma("identity", size = 10, dispersion = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+})
+
+
+testthat::skip_on_cran()
+test_that("sampling works (vinverse.gaussian)", {
+  set.seed(42)
+  n_obs <- 200L
+  W <- generateW("rectangle", 100, 2, 10)
+  model_orders <- list(intercept = "homogeneous", past_obs = 2, past_mean = 1)
+  parameter <- list(intercept = 0.5, past_obs = matrix(c(0.3, 0.2, 0.05), nrow = 3), 
+                    past_mean = matrix(c(0.1, 0.05), nrow = 2))
+  
+  # 1/mu^2-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vinverse.gaussian("1/mu^2", size = 10, dispersion = 2, const = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vinverse.gaussian("1/mu^2", size = 10, dispersion = 2, const = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+  # identity-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vinverse.gaussian("identity", size = 10, dispersion = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vinverse.gaussian("identity", size = 10, dispersion = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+ # log link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vinverse.gaussian("log", size = 10, dispersion = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vinverse.gaussian("log", size = 10, dispersion = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+})
+
+
+testthat::skip_on_cran()
+test_that("sampling works (vnormal)", {
+  set.seed(42)
+  n_obs <- 200L
+  W <- generateW("rectangle", 100, 2, 10)
+  model_orders <- list(intercept = "homogeneous", past_obs = 2, past_mean = 1)
+  parameter <- list(intercept = 0.5, past_obs = matrix(c(0.3, 0.2, 0.05), nrow = 3), 
+                    past_mean = matrix(c(0.1, 0.05), nrow = 2))
+  
+  # log-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnormal("log", size = 10, dispersion = 2, const = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnormal("log", size = 10, dispersion = 2, const = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+  # identity-link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnormal("identity", size = 10, dispersion = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnormal("identity", size = 10, dispersion = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+
+ # inverse link
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnormal("inverse", size = 10, dispersion = 2))
+  expect_true(is.list(sim))
+  sim <- glmstarma.sim(n_obs, parameter, model_orders, W, family = vnormal("inverse", size = 10, dispersion = 2, copula = "frank", copula_param = 2))
+  expect_true(is.list(sim))
+})